Octree and Barnes Hut Implementation (#20)
* made progress on octree class; working on building the octree class

* finished implementing octree class and members/functions; need to fix segmentation fault in testing

* update testing and previous implementation; debug octree implementation and makefile configuration

* add inclusion of <memory> to octree files

* add tests for body implementation; slightly edit body.cpp implementation

* fixed implementation of octree

* fixed octree insertion implementation; added testing for octree insert, but need more coverage

* fixed basic pointers vs shared pointers in octree implementation; added an envOctree member to the Grav Environment class; added some tests

* fixed issue with types; need to add more coverage and implement barnes-hut algo walking down the tree

* change every double to a float; implemented barnes-hut force algo; need to test

* implemented barnes-hut force calculation; need to implement testing

* Add functional header

* added testing for barnes-hut force calculation algorithm

* add an include header of math.h to test_environment

* fix type issue in test of barnes-hut algo

* added more testing for octree class; changed totalMass to a float from a pointer

* added line to workaround makefile issue on John's laptop; changed to floats from doubles; implemented work with Adam's config stuff

* change include to INC_DIR variable; fix warning in test_statistics.cpp

* small change to push and test workflow

* commented out test case that detects for errors in loading file

* add line to coverage to debug

* added debugging print to coverage workflow

* added dummy executable for the statistics file; changed += from Makefile; added download of bc in coverage workflow

* adjut Makefile for John's local machine; testing automation on github

* Fix small typo in Makefile to fix automation (LDLFLAGS --> LDFLAGS)

* change initilization of coordinate looping in environment to at first, position of first pointer

---------

+++ include/particle.h
@@ -7,13 +7,13 @@ class Particle {
     
     public:
         // Constructors
-        Particle(const std::array<double, 3>* initial_position, const std::array<double, 3>* initial_velocity, double mass);
+        Particle(const std::array<float, 3>* initial_position, const std::array<float, 3>* initial_velocity, float mass);
 
         // Define member functions
-        void update (const std::array<double, 3>* force, const double timestep);
+        void update (const std::array<float, 3>* force, const float timestep);
 
         // Instantiation of the physical members
-        std::array<double, 3> position;
-        std::array<double, 3> velocity;
-        double mass;
+        std::array<float, 3> position;
+        std::array<float, 3> velocity;
+        float mass;
 };
